% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{av_patient_tumour_merge}
\alias{av_patient_tumour_merge}
\title{Merge of patient and tumour df}
\usage{
av_patient_tumour_merge(df1 = sim_av_patient, df2 = sim_av_tumour)
}
\arguments{
\item{df1}{a dataframe to be merged with df2. df1 is default set to `sim_av_patient`.}

\item{df2}{a dataframe to be merged with df1. df2 is default set to `sim_av_tumour`.}
}
\value{
a merged dataframe
}
\description{
A simple merge function for gathering the two dataframes ....
}
