% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{cancer_grouping}
\alias{cancer_grouping}
\title{Group Cancer Diagnoses by ICD-10 Codes}
\usage{
cancer_grouping(df)
}
\arguments{
\item{df}{A data frame containing a column `SITE_ICD10_O2_3CHAR` with ICD-10 codes.}
}
\value{
A data frame with an added column `diag_group` for the cancer category.
}
\description{
Groups cancer diagnoses into broader categories based on ICD-10 codes.
}
\details{
This function aggregates ICD-10 cancer codes into groups 
to simplify analyses. It covers common cancer groups and provides a default 
"Ill-defined and unspecified" category for unmatched codes.

The ICD-10 ranges are derived from NHS guidelines. See the source for details:
<https://digital.nhs.uk/ndrs/data/data-outputs/cancer-data-hub/cancer-prevalence>.
}
\examples{
sim_av_tumour <- cancer_grouping(sim_av_tumour)
}
