% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir}
\alias{create_dir}
\title{Create a new directory if it does not exist.}
\usage{
create_dir(dir_path = "Outputs", verbose = TRUE)
}
\arguments{
\item{dir_path}{A character string defining a path of the directory to create. Defaults is "Outputs".}

\item{verbose}{A logical value to decide if there should be a printed messages about the directory creation process. Defaults is TRUE.}
}
\value{
invisible return of the `dir_path`
}
\description{
This function creates a directory at the specified path if it does not already exist. The default name is "Outputs". To illustrate the folder is meant for various outputs of the package.
There is also optional verbose output to inform the user about the action taken.
}
