% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir_if_none}
\alias{create_dir_if_none}
\title{Creating a Dir if there is None}
\usage{
create_dir_if_none(dir)
}
\arguments{
\item{dir}{the direction to the folder}
}
\value{
folder
}
\description{
Creating a Dir if there is None
}
