% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_management.R
\name{end_time}
\alias{end_time}
\title{Record the Current End Time}
\usage{
end_time()
}
\value{
A `POSIXct` object representing the current system time.
}
\description{
Captures the current system time to be used as the end time for timing execution.
}
\examples{
end_time() 
}
