% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extended_summary}
\alias{extended_summary}
\title{Generate an Extended Summary of a Data Frame}
\usage{
extended_summary(df)
}
\arguments{
\item{df}{A data frame to summarize.}
}
\value{
A data frame summarizing:
- Column names
- Number of missing values
- Number of unique values
- Data class of each column
}
\description{
Summarizes a data frame, including missing values, unique values, and column classes.
}
\examples{
summary_df <- extended_summary(sim_av_patient)
}
