% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{group_ethnicity}
\alias{group_ethnicity}
\title{Group Ethnicity Categories}
\usage{
group_ethnicity(df)
}
\arguments{
\item{df}{A data frame containing an `ETHNICITY` column with ethnicity codes.}
}
\value{
A data frame with an added column `Grouped_Ethinicity`.
}
\description{
Maps ethnicity codes to broader categories.
}
\details{
A function which can be used to map the 24 ethnicity code into 5 categories. The mapping for is based on the guide from NHS which is also the dataholders of the CAS database.
The source for the mapping can be found on this link: https://digital.nhs.uk/data-and-information/data-collections-and-data-sets/data-sets/mental-health-services-data-set/submit-data/data-quality-of-protected-characteristics-and-other-vulnerable-groups/ethnicity
}
