% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_func}
\alias{log_func}
\title{System logger and timer logger}
\usage{
log_func(func, log_file = "detailed_log.txt")
}
\arguments{
\item{func}{The function to be logged.}

\item{log_file}{The name of the log file.}

\item{verbose}{If TRUE, prints log messages to the console.}

\item{log_level}{The log level to use (e.g., "INFO", "WARN", "ERROR").}
}
\value{
The result of the function `func`.
}
\description{
This function logs the execution time and system information of a given function, using the `logger` package for flexible logging.
}
