% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{System logger and timer logger}
\usage{
logger()
}
\arguments{
\item{func}{The function to be logged.}

\item{...}{Additional arguments to be passed to the `func`.}
}
\value{
The return value of the `func`.
}
\description{
This function logs the execution time and system information of a given function. It's designed to be used as a wrapper around other functions to track performance and identify potential bottlenecks.
}
\examples{
# Example function
my_function <- function(x) {
  Sys.sleep(2)
  return(x^2)
}

# Log the execution of my_function
result <- logger(my_function, 5)
print(result)

}
