% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_simulacrum_request}
\alias{open_simulacrum_request}
\title{Open Simulacrum Data Request Page}
\usage{
open_simulacrum_request()
}
\description{
Opens the Simulacrum data download webpage, checks if it can be reached and prints instructions.
}
\examples{
open_simulacrum_request()
}
