% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_col2.R
\name{read_csv}
\alias{read_csv}
\title{Read and Load Multiple CSV Files}
\usage{
read_csv(dir, selected_files = NULL)
}
\arguments{
\item{dir}{A character string specifying the directory containing the CSV files.}

\item{selected_files}{A character vector of file names (without extensions) to load. If NULL, all required files are loaded. Default is NULL.}
}
\value{
A named list where each element is a data frame corresponding to a loaded CSV file. The names of the list elements are the file names without extensions.
}
\description{
This function reads multiple CSV files from a specified directory. It validates the presence of required files, allows optional selection of specific files to load, and tracks progress during the loading process.
}
\details{
The function takes a directory path as input, where the CSV files downloaded from Simulacrum are located. 
It checks if the files are valid CSVs and verifies that all required CSV files are present in the directory by comparing the list of files in the directory with an internal list (required_files).
Optionally, users can specify certain CSV files to load using the selected_files argument.

Once the files are selected, they are imported into a list of data frames. Each data frame corresponds to a CSV file, and users can assign these data frames to variables as needed. 
It is recommended to name the data frames after their corresponding CSV files for clarity.

To enhance user experience, the function uses the pbapply package to display a progress bar and track the time taken to import each file.
}
\examples{
# Load all required files
data_frames <- read_csv("directory/to/csv")
sim_av_patient <- data_frames$sim_av_patient

# Load specific files
data_frames <- read_csv("directory/to/csv"), selected_files = c("sim_av_patient", "sim_av_tumour"))

}
