% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_queries.R
\name{sql_full_av}
\alias{sql_full_av}
\title{Generate SQL Query for Full Joins on small series of Tables}
\usage{
sql_full_av()
}
\value{
The SQL query as a printed string.
}
\description{
These functions generate SQL queries for joining multiple tables within the categories of the Simulacrum, like av, rtds and sact. Also a full join of all the tables:
}
\examples{
sql_full_av()
sql_full_sact()
sql_full_rtds()
sql_full()
}
