% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{sql_sim_av_patient}
\alias{sql_sim_av_patient}
\title{Generate SQL Query for Individual Tables}
\usage{
sql_sim_av_patient()
}
\value{
The SQL query as a printed string.
}
\description{
Each of these functions generates an SQL query for selecting all rows from a certain table in the Simulacrum dataset.
}
\examples{
sql_sim_av_patient()
sql_av_tumour_patient()
sql_sim_av_gene()
}
