% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_management.R
\name{start_time}
\alias{start_time}
\title{Record the Current Start Time}
\usage{
start_time()
}
\value{
A `POSIXct` object representing the current system time.
}
\description{
Captures the current system time to be used as the start time for timing execution.
}
\examples{
start <- current_start_time()
}
