% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{survival_days}
\alias{survival_days}
\title{Calculate Survival Days}
\usage{
survival_days(df)
}
\arguments{
\item{df}{A data frame containing the necessary columns: 'DIAGNOSISDATEBEST', 
'VITALSTATUSDATE', and 'VITALSTATUS'.}
}
\value{
The input data frame with an additional column `date_diff`, 
representing the number of days between `DIAGNOSISDATEBEST` and `VITALSTATUSDATE`.
Also, a column called `date_to_death` which is the same column as `date_diff` but where the column `VITALSTATUS` is set to be equal to `D` which means death.
}
\description{
This function calculates the number of days between the diagnosis date 
and the vital status date for each row in the data frame.
}
\details{
This function helps the researchers who wants to survival analysis, by offering a fast a custombuilt function to compute the survial days between diagnosis day and day of death
The functions does ask the user to merge the two dataframe `sim_av_patient` and `sim_av_tumour` before the function can be used as the variables needed for the operations are split in between the two tables.
}
